SNOW-115108: fix arrow data not allocating more than 2GB

+++ src/test/java/net/snowflake/client/core/arrow/BitToBooleanConverterTest.java
@@ -25,7 +25,7 @@ public class BitToBooleanConverterTest extends BaseConverterTest
   /**
    * allocator for arrow
    */
-  private BufferAllocator allocator = new RootAllocator(Integer.MAX_VALUE);
+  private BufferAllocator allocator = new RootAllocator(Long.MAX_VALUE);
 
   private Random random = new Random();
